t1x2y1/requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.25
python-dotenv==1.0.0
pydantic==2.5.3
psycopg2-binary==2.9.9
python-logging-loki==0.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
ratelimit==2.2.1
aiohttp==3.9.1
asyncio==3.4.3
python-dateutil==2.8.2
aiosqlite==0.19.0
async-timeout==4.0.3
asyncio-throttle==0.1.0
